fix: fold model wiring files to satisfy pkg-file-count gate
Merge model service host wiring into model_catalog.go and fold new API,
CLI, and service tests into existing package files so pkg/api, pkg/cli/run,
and pkg/service stay within the 15-file limit without losing coverage.
